+++ NEWS.md
@@ -1,3 +1,8 @@
+# ICIKendallTau 0.3.0
+
+- Handling parallel execution differently to avoid really large matrix issues on each core.
+- Introduces the `return_matrix` parameter to `ici_kendalltau` that allows to return the results in the form of a `data.frame` instead of lists of matrices.
+
 # ICIKendalltau 0.2.8
 
 - Made all error outputs the same length as the default output but containing `NA`.